re organize directory to make more sense
