+++ node/src/main/scala/coop/rchain/node/revVaultExport/RhoTrieTraverser.scala
@@ -1,22 +1,16 @@
-package coop.rchain.casper.util.rholang
+package coop.rchain.node.revVaultExport
 
 import cats.effect.Sync
 import cats.implicits._
 import com.google.protobuf.ByteString
 import coop.rchain.casper.genesis.contracts.StandardDeploys
+import coop.rchain.casper.util.rholang.Tools
 import coop.rchain.crypto.hash.Keccak256
-import coop.rchain.models.Expr.ExprInstance.{
-  EListBody,
-  EMapBody,
-  ETupleBody,
-  GByteArray,
-  GInt,
-  GString
-}
+import coop.rchain.models.Expr.ExprInstance._
 import coop.rchain.models.GUnforgeable.UnfInstance.GPrivateBody
-import coop.rchain.models.{EList, ETuple, Expr, GPrivate, GUnforgeable, Par, ParMap}
-import coop.rchain.rholang.interpreter.{RhoRuntime, RhoType}
+import coop.rchain.models._
 import coop.rchain.rholang.interpreter.storage.serializePar
+import coop.rchain.rholang.interpreter.{RhoRuntime, RhoType}
 import coop.rchain.shared.Serialize
 
 import scala.annotation.tailrec
