+++ node/src/main/scala/coop/rchain/node/revVaultExport/StateBalances.scala
@@ -1,4 +1,4 @@
-package coop.rchain.node.balance
+package coop.rchain.node.revVaultExport
 
 import cats.Parallel
 import cats.effect.{Concurrent, ContextShift}
@@ -7,14 +7,13 @@ import com.google.protobuf.ByteString
 import coop.rchain.blockstorage.KeyValueBlockStore
 import coop.rchain.casper.storage.RNodeKeyValueStoreManager
 import coop.rchain.casper.storage.RNodeKeyValueStoreManager.legacyRSpacePathPrefix
-import coop.rchain.casper.util.rholang.VaultBalanceGetter
 import coop.rchain.crypto.codec.Base16
 import coop.rchain.metrics.{Metrics, NoopSpan}
 import coop.rchain.models.{BindPattern, ListParWithRandom, Par, TaggedContinuation}
 import coop.rchain.rholang.interpreter.RhoRuntime
+import coop.rchain.rspace.hashing.Blake2b256Hash
 import coop.rchain.rspace.syntax._
 import coop.rchain.rspace.{Match, RSpace}
-import coop.rchain.rspace.hashing.Blake2b256Hash
 import coop.rchain.shared.Log
 
 import java.nio.file.{Files, Path}
