+++ node/src/main/scala/coop/rchain/node/revVaultExport/VaultBalanceGetter.scala
@@ -1,4 +1,4 @@
-package coop.rchain.casper.util.rholang
+package coop.rchain.node.revVaultExport
 
 import cats.effect.Sync
 import cats.implicits._
@@ -6,8 +6,8 @@ import com.google.protobuf.ByteString
 import coop.rchain.crypto.codec.Base16
 import coop.rchain.crypto.hash.Blake2b512Random
 import coop.rchain.models.Expr.ExprInstance.{GInt, GString}
-import coop.rchain.models.{Expr, GPrivate, Par, Send}
 import coop.rchain.models.rholang.implicits._
+import coop.rchain.models.{Expr, GPrivate, Par, Send}
 import coop.rchain.rholang.interpreter.RhoRuntime
 import coop.rchain.rholang.interpreter.accounting.Cost
