+++ node/src/main/scala/coop/rchain/node/revVaultExport/mainNet1/StateBalanceMain.scala
@@ -1,7 +1,8 @@
-package coop.rchain.node.balance
+package coop.rchain.node.revVaultExport.mainNet1
 
 import cats.effect._
 import coop.rchain.crypto.codec.Base16
+import coop.rchain.node.revVaultExport.StateBalances
 import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
 import org.rogach.scallop.ScallopConf
