+++ node/src/main/scala/coop/rchain/node/revVaultExport/mainNet1/reporting/MergeBalanceMain.scala
@@ -1,4 +1,4 @@
-package coop.rchain.node.balance
+package coop.rchain.node.revVaultExport.mainNet1.reporting
 
 import cats.effect.Sync
 import cats.implicits._
@@ -6,25 +6,21 @@ import com.google.protobuf.ByteString
 import coop.rchain.blockstorage.KeyValueBlockStore
 import coop.rchain.casper.storage.RNodeKeyValueStoreManager
 import coop.rchain.casper.storage.RNodeKeyValueStoreManager.legacyRSpacePathPrefix
-import coop.rchain.rholang.interpreter.RhoRuntime
-import org.rogach.scallop.ScallopConf
 import coop.rchain.casper.syntax._
-import coop.rchain.rspace.syntax._
-import coop.rchain.blockstorage.syntax._
-import coop.rchain.shared.syntax._
 import coop.rchain.crypto.codec.Base16
 import coop.rchain.metrics.{Metrics, NoopSpan, Span}
 import coop.rchain.models.{BindPattern, ListParWithRandom, Par, TaggedContinuation}
+import coop.rchain.rholang.interpreter.RhoRuntime
+import coop.rchain.rspace.syntax._
 import coop.rchain.rspace.{Match, RSpace}
-import coop.rchain.rspace.hashing.Blake2b256Hash
 import coop.rchain.shared.Log
 import monix.eval.Task
+import monix.execution.Scheduler.Implicits.global
+import org.rogach.scallop.ScallopConf
 
 import java.io.PrintWriter
 import java.nio.file.{Files, Path}
-import scala.collection.concurrent.TrieMap
 import scala.io._
-import monix.execution.Scheduler.Implicits.global
 
 /**
   * The `merge-balance-main` would generate a csv file and the format is like below
