+++ node/src/main/scala/coop/rchain/node/revVaultExport/mainNet1/reporting/TransactionBalanceMain.scala
@@ -1,7 +1,7 @@
-package coop.rchain.node.balance
+package coop.rchain.node.revVaultExport.mainNet1.reporting
 
 import cats.effect._
-import coop.rchain.crypto.codec.Base16
+import coop.rchain.node.revVaultExport.reporting.TransactionBalances
 import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
 import org.rogach.scallop.ScallopConf
