+++ node/src/main/scala/coop/rchain/node/revVaultExport/reporting/Transaction.scala
@@ -1,16 +1,16 @@
-package coop.rchain.node.balance
+package coop.rchain.node.revVaultExport.reporting
 
 import cats.effect.Concurrent
 import cats.implicits._
 import coop.rchain.shared.Log
+import coop.rchain.shared.syntax._
 import coop.rchain.store.{KeyValueTypedStore, LmdbStoreManager}
 import io.circe.Decoder.Result
+import io.circe.generic.auto._
 import io.circe.generic.semiauto.deriveDecoder
-import io.circe.{Decoder, HCursor, Json}
 import io.circe.parser._
-import io.circe._
-import io.circe.generic.auto._, io.circe.syntax._
-import coop.rchain.shared.syntax._
+import io.circe.syntax._
+import io.circe.{Decoder, HCursor, Json, _}
 import scodec.bits.BitVector
 import scodec.{Attempt, Codec, DecodeResult, Err, SizeBound}
