+++ node/src/main/scala/coop/rchain/node/revVaultExport/reporting/TransactionBalances.scala
@@ -1,31 +1,32 @@
-package coop.rchain.node.balance
+package coop.rchain.node.revVaultExport.reporting
 
 import cats.Parallel
-import cats.implicits._
 import cats.effect.{Concurrent, ContextShift, Sync}
+import cats.implicits._
 import cats.mtl.FunctorRaise
 import com.google.protobuf.ByteString
-import coop.rchain.blockstorage.{BlockStore, KeyValueBlockStore}
 import coop.rchain.blockstorage.dag.{BlockDagKeyValueStorage, BlockDagRepresentation}
-import coop.rchain.casper.storage.RNodeKeyValueStoreManager
-import coop.rchain.casper.storage.RNodeKeyValueStoreManager.legacyRSpacePathPrefix
-import coop.rchain.metrics.{Metrics, NoopSpan, Span}
-import coop.rchain.rholang.interpreter.util.RevAddress
-import coop.rchain.crypto.codec.Base16
-import coop.rchain.models.{BindPattern, ListParWithRandom, Par, TaggedContinuation}
-import coop.rchain.rholang.interpreter.RhoRuntime
-import coop.rchain.rspace.{Match, RSpace}
-import coop.rchain.rspace.syntax._
-import coop.rchain.casper.syntax._
 import coop.rchain.blockstorage.util.io.IOError
+import coop.rchain.blockstorage.{BlockStore, KeyValueBlockStore}
 import coop.rchain.casper.PrettyPrinter
 import coop.rchain.casper.genesis.contracts.StandardDeploys
 import coop.rchain.casper.protocol.BlockMessage
-import coop.rchain.casper.util.rholang.RhoTrieTraverser
+import coop.rchain.casper.storage.RNodeKeyValueStoreManager
+import coop.rchain.casper.storage.RNodeKeyValueStoreManager.legacyRSpacePathPrefix
+import coop.rchain.casper.syntax._
 import coop.rchain.casper.util.{BondsParser, VaultParser}
 import coop.rchain.crypto.PrivateKey
+import coop.rchain.crypto.codec.Base16
 import coop.rchain.crypto.signatures.Secp256k1
-import coop.rchain.node.balance.SpecialCase.getSpecialTransfer
+import coop.rchain.metrics.{Metrics, NoopSpan, Span}
+import coop.rchain.models.{BindPattern, ListParWithRandom, Par, TaggedContinuation}
+import coop.rchain.node.revVaultExport.mainNet1.reporting.SpecialCase.getSpecialTransfer
+import coop.rchain.node.revVaultExport.RhoTrieTraverser
+import coop.rchain.node.revVaultExport.mainNet1.reporting.PerValidatorVaults
+import coop.rchain.rholang.interpreter.RhoRuntime
+import coop.rchain.rholang.interpreter.util.RevAddress
+import coop.rchain.rspace.syntax._
+import coop.rchain.rspace.{Match, RSpace}
 import coop.rchain.shared.Log
 import coop.rchain.shared.syntax._
 import fs2._
